put validation in deathmatch command instead of manager and fix styling
