+++ javascript/features/death_match/death_match_commands.js
@@ -9,7 +9,7 @@ import { Menu } from 'components/menu/menu.js';
 // Contains the /dm command for players to allow to teleport to a DM zone with specific weapons.
 export class DeathMatchCommands {    
 
-    constructor(manager) {
+    constructor(abuse, manager) {
         server.commandManager.buildCommand('deathmatch')
             .parameters([{ name: 'zone', type: CommandBuilder.NUMBER_PARAMETER, optional: true }])
             .sub('leave')
@@ -19,22 +19,38 @@ export class DeathMatchCommands {
     
         .build(DeathMatchCommands.prototype.onDmCommand.bind(this));
 
-        this.manager_ = manager;
+        this.manager_ = manager;        
+        this.abuse_ = abuse;
     }
 
     // The death match command is being used.
     async onDmCommand(player, zone) { 
-        if(isNaN(zone)) {
+        if(isNaN(zone) || !DeathMatchLocation.hasLocation(zone)) {
             const dialog = new Menu('Choose a death match zone.', ['Zone', 'Name', 'Teams', 'Lag shot']);
             for(const zoneId of this.manager_.validDmZones()) {
                 const location = DeathMatchLocation.getById(zoneId);  
-                dialog.addItem(location.id, location.name, location.hasTeams, location.lagShot, 
+                dialog.addItem(location.id, location.name, location.hasTeams ? "Yes" : "No", 
+                                location.lagShot ? "Yes": "No", 
                     (player) => {
                         this.manager_.goToDmZone(player, zoneId);
                     });
             }
             
             await dialog.displayForPlayer(player);
+            return;
+        }
+        
+        if (player.activity !== Player.PLAYER_ACTIVITY_JS_DM_ZONE &&
+            player.activity !== Player.PLAYER_ACTIVITY_NONE) {
+            player.sendMessage(Message.DEATH_MATCH_TELEPORT_BLOCKED, "you are in another activity.");
+            return;
+        }
+
+        const teleportStatus = this.abuse_().canTeleport(player, { enforceTimeLimit: true });
+
+        // Bail out if the |player| is not currently allowed to teleport.
+        if (!teleportStatus.allowed) {
+            player.sendMessage(Message.DEATH_MATCH_TELEPORT_BLOCKED, teleportStatus.reason);
             return;
         }
