+++ javascript/features/death_match/death_match_commands.test.js
@@ -2,6 +2,7 @@
 // Use of this source code is governed by the MIT license, a copy of which can
 // be found in the LICENSE file.
 
+import AbuseConstants from 'features/abuse/abuse_constants.js';
 import DeathMatch from "features/death_match/death_match.js";
 
 describe('DeathMatchManager', (it, beforeEach) => {
@@ -20,26 +21,32 @@ describe('DeathMatchManager', (it, beforeEach) => {
         manager = deathMatch.manager_;
     });
 
-    it('should allow to use gang zone 1', async(assert) => {
+    it('should not enable players to go to a DM zone when they might abuse it', async (assert) => {
         const gunther = server.playerManager.getById(0 /* Gunther */);
+        const russell = server.playerManager.getById(1 /* Russell */);
+
+        gunther.identify({ userId: 42 });
+        gunther.shoot({ target: russell });
 
         assert.isTrue(await gunther.issueCommand('/deathmatch 1'));
 
-        assert.equal(gunther.messages.length, 3);
-        assert.includes(gunther.messages[0], 
-            Message.format(Message.DEATH_MATCH_INSTRUCTION_LEAVE, 0));
+        assert.equal(gunther.messages.length, 1);
+        assert.equal(
+            gunther.messages[0],
+            Message.format(Message.DEATH_MATCH_TELEPORT_BLOCKED,
+                AbuseConstants.REASON_FIRED_WEAPON));
     });
 
-    it('should show message for player if using invalid dm zone', async(assert) => {
+    it('should not allow players to go to a death match if they are in another activity', async (assert) => {
         const gunther = server.playerManager.getById(0 /* Gunther */);
+        gunther.activity = Player.PLAYER_ACTIVITY_JS_RACE;
+       
+        assert.isTrue(await gunther.issueCommand('/deathmatch 1'));
 
-        assert.isTrue(await gunther.issueCommand('/deathmatch 0'));
-
-        assert.equal(gunther.messages.length, 2);
-        assert.includes(gunther.messages[0], 
-            Message.format(Message.DEATH_MATCH_INVALID_ZONE, 0));
-        assert.includes(gunther.messages[1], Message.format(Message.DEATH_MATCH_AVAILABLE_ZONES, 
-            manager.validDmZones().join(', ')));
+        assert.equal(gunther.messages.length, 1);
+        assert.equal(
+            gunther.messages[0],
+            Message.format(Message.DEATH_MATCH_TELEPORT_BLOCKED, "you are in another activity."));
     });
 
     it('should allow to use death match leave', async(assert) => {
@@ -54,6 +61,28 @@ describe('DeathMatchManager', (it, beforeEach) => {
         gunther.respondToDialog({ listitem: 0 /* Assumed `zone 1` */ });
 
         assert.isTrue(await gunther.issueCommand('/deathmatch'));
+
+        assert.equal(gunther.messages.length, 3);
+        assert.includes(gunther.messages[0], 
+            Message.format(Message.DEATH_MATCH_INSTRUCTION_LEAVE, 0));
+    });
+
+    it('should show message for player if using invalid dm zone', async(assert) => {
+        const gunther = server.playerManager.getById(0 /* Gunther */);        
+
+        gunther.respondToDialog({ listitem: 0 /* Assumed `zone 1` */ });
+
+        assert.isTrue(await gunther.issueCommand('/deathmatch 0'));
+
+        assert.equal(gunther.messages.length, 3);
+        assert.includes(gunther.messages[0], 
+            Message.format(Message.DEATH_MATCH_INSTRUCTION_LEAVE, 0));
+    });
+
+    it('should allow to use gang zone 1', async(assert) => {
+        const gunther = server.playerManager.getById(0 /* Gunther */);
+
+        assert.isTrue(await gunther.issueCommand('/deathmatch 1'));
 
         assert.equal(gunther.messages.length, 3);
         assert.includes(gunther.messages[0],
